In progress.

+++ gulp/tasks/generate-definitions.js
@@ -2,51 +2,61 @@ var gulp = require("gulp");
 var fs = require("fs");
 var path = require("path");
 var map = require("gulp-map");
+var glob = require("glob");
 
 gulp.task(
     "generate-definitions",
-    async (cb) => {
+    async () => {
 
-        // console.log("START! generate-definitions.js");
+        await new Promise(
+            (resolve) => {
 
-        var getSafeDirPath = function(dirPath) {
-            dirPath += dirPath.charAt(dirPath.length - 1) == "/" ? "" : "/";
-            return dirPath;
-        };
+                // console.log("START! generate-definitions.js");
 
-        var argv = {};
+                var getSafeDirPath = function(dirPath) {
+                    dirPath += dirPath.charAt(dirPath.length - 1) == "/" ? "" : "/";
+                    return dirPath;
+                };
 
-        var basePath = "./src/";
-        // console.log("BEFORE argv.src: " + argv.src);
-        argv.src = argv.src ? argv.src : basePath;
-        // Adding a closing slash to make correct folder path (if needed)
-        argv.src = getSafeDirPath(argv.src);
-        // console.log("AFTER argv.src: " + argv.src);
+                var argv = {};
 
-        // console.log("BEFORE argv.outFile: " + argv.outFile);
-        argv.outFile = argv.outFile ? argv.outFile : "index";
-        // console.log("AFTER argv.outFile: " + argv.outFile);
+                var basePath = "./src/";
+                // console.log("BEFORE argv.src: " + argv.src);
+                argv.src = argv.src ? argv.src : basePath;
+                // Adding a closing slash to make correct folder path (if needed)
+                argv.src = getSafeDirPath(argv.src);
+                // console.log("AFTER argv.src: " + argv.src);
 
-        // console.log("BEFORE argv.outDir: " + argv.outDir);
-        argv.outDir = argv.outDir ? argv.outDir : basePath;
-        argv.outDir = getSafeDirPath(argv.outDir);
-        // console.log("AFTER argv.outDir: " + argv.outDir);
+                // console.log("BEFORE argv.outFile: " + argv.outFile);
+                argv.outFile = argv.outFile ? argv.outFile : "index";
+                // console.log("AFTER argv.outFile: " + argv.outFile);
 
-        var outFileName = argv.outFile + ".ts";
-        // console.log("outFileName: " + outFileName);
-        // Remove prev index file
-        await fs.unlinkSync(basePath + outFileName);
+                // console.log("BEFORE argv.outDir: " + argv.outDir);
+                argv.outDir = argv.outDir ? argv.outDir : basePath;
+                argv.outDir = getSafeDirPath(argv.outDir);
+                // console.log("AFTER argv.outDir: " + argv.outDir);
 
-        var resultDeclarationText = "";
+                var outFileName = argv.outFile + ".ts";
+                // console.log("outFileName: " + outFileName);
+                // Remove prev index file
+                if (fs.existsSync(basePath + outFileName)) {
+                    fs.unlinkSync(basePath + outFileName);
+                }
 
-        // console.log("Imported files:");
-        var tempSettings = [argv.src + "**/*.ts", "!./!**/!*.d.ts"];
+                var resultDeclarationText = "";
 
-        gulp.src(tempSettings)
-                .pipe(
-                    map(
-                        (file) => {
-                            let importPath = path.relative(basePath, file.path);
+                // console.log("Imported files:");
+                var tempSettings = `${argv.src}**/*.ts`;
+
+                glob(
+                    tempSettings,
+                    (error, files) => {
+                        var filesCount = files.length;
+                        for (var fileIndex = 0; fileIndex < filesCount; fileIndex++) {
+                            var singleFile = files[fileIndex];
+                            console.log("singleFile: ", singleFile);
+
+                            let importPath = path.relative(basePath, singleFile);
                             if (importPath.indexOf(".d.ts") != -1) {
                                 importPath = importPath.substr(0, importPath.lastIndexOf(".d.ts"));
                             } else if (importPath.indexOf(".ts") != -1) {
@@ -57,11 +67,7 @@ gulp.task(
                             resultDeclarationText += "export * from '" + "./" + importPath + "'";
                             resultDeclarationText += "\n";
                         }
-                    )
-                )
-                .on(
-                    "end",
-                    () => {
+
                         console.log("Declarations: ");
                         console.log(resultDeclarationText);
 
@@ -70,10 +76,12 @@ gulp.task(
                             resultDeclarationText,
                             () => {
                                 console.log("WRITE COMPLETE!");
-                                cb();
+                                resolve();
                             }
                         );
                     }
                 );
+            }
+        );
     }
 );
